users: rename go mod
