+++ apps/users/api-server/create_user.go
@@ -3,7 +3,7 @@ package main
 import (
 	"encoding/json"
 
-	"github.com/SunSince90/ship-krew/users/api-server/pkg/api"
+	"github.com/SunSince90/ship-krew/apps/users/api-server/pkg/api"
 	fiber "github.com/gofiber/fiber/v2"
 	uuid "github.com/satori/go.uuid"
 )
