+++ apps/users/api-server/delete_user.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/SunSince90/ship-krew/users/api-server/pkg/api"
+	"github.com/SunSince90/ship-krew/apps/users/api-server/pkg/api"
 	fiber "github.com/gofiber/fiber/v2"
 )
