+++ apps/users/api-server/get_user.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/SunSince90/ship-krew/users/api-server/pkg/api"
+	"github.com/SunSince90/ship-krew/apps/users/api-server/pkg/api"
 	fiber "github.com/gofiber/fiber/v2"
 )
