+++ apps/users/api-server/list_users.go
@@ -3,7 +3,7 @@ package main
 import (
 	"strconv"
 
-	"github.com/SunSince90/ship-krew/users/api-server/pkg/api"
+	"github.com/SunSince90/ship-krew/apps/users/api-server/pkg/api"
 	fiber "github.com/gofiber/fiber/v2"
 )
